#55 Change float to int for heuristic

+++ bachelor/sim/src/ai/path_finding.cpp
@@ -10,7 +10,7 @@ glm::ivec2 world_to_grid(glm::vec2 pos, int grid)
     return {static_cast<int>(pos.x) / static_cast<int>(grid), static_cast<int>(pos.y) / static_cast<int>(grid)};
 }
 
-float path_heuristic(PathGrid start, PathGrid goal)
+int path_heuristic(PathGrid start, PathGrid goal)
 {
     auto p = start - goal;
     return p.x * p.x + p.y * p.y;
